+++ librelane/scripts/openroad/common/pdn_cfg.tcl
@@ -1,3 +1,7 @@
+# Copyright 2025 LibreLane Contributors
+#
+# Adapted from OpenLane
+#
 # Copyright 2020-2022 Efabless Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,6 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+source $::env(SCRIPTS_DIR)/openroad/common/io.tcl
 source $::env(SCRIPTS_DIR)/openroad/common/set_global_connections.tcl
 set_global_connections
 
@@ -43,12 +48,24 @@ foreach vdd $::env(VDD_NETS) gnd $::env(GND_NETS) {
 set_voltage_domain -name CORE -power $::env(VDD_NET) -ground $::env(GND_NET) \
     -secondary_power $secondary
 
+
+
 if { $::env(PDN_MULTILAYER) == 1 } {
+
+    set arg_list [list]
+    if { $::env(PDN_ENABLE_PINS) } {
+        lappend arg_list -pins "$::env(PDN_VERTICAL_LAYER) $::env(PDN_HORIZONTAL_LAYER)"
+    }
+
     define_pdn_grid \
         -name stdcell_grid \
         -starts_with POWER \
         -voltage_domain CORE \
-        -pins "$::env(PDN_VERTICAL_LAYER) $::env(PDN_HORIZONTAL_LAYER)"
+        {*}$arg_list
+
+    set arg_list [list]
+    append_if_equals arg_list PDN_EXTEND_TO "core_ring" -extend_to_core_ring
+    append_if_equals arg_list PDN_EXTEND_TO "boundary" -extend_to_boundary
 
     add_pdn_stripe \
         -grid stdcell_grid \
@@ -57,7 +74,8 @@ if { $::env(PDN_MULTILAYER) == 1 } {
         -pitch $::env(PDN_VPITCH) \
         -offset $::env(PDN_VOFFSET) \
         -spacing $::env(PDN_VSPACING) \
-        -starts_with POWER -extend_to_core_ring
+        -starts_with POWER \
+        {*}$arg_list
 
     add_pdn_stripe \
         -grid stdcell_grid \
@@ -66,17 +84,28 @@ if { $::env(PDN_MULTILAYER) == 1 } {
         -pitch $::env(PDN_HPITCH) \
         -offset $::env(PDN_HOFFSET) \
         -spacing $::env(PDN_HSPACING) \
-        -starts_with POWER -extend_to_core_ring
+        -starts_with POWER \
+        {*}$arg_list
 
     add_pdn_connect \
         -grid stdcell_grid \
         -layers "$::env(PDN_VERTICAL_LAYER) $::env(PDN_HORIZONTAL_LAYER)"
 } else {
+
+    set arg_list [list]
+    if { $::env(PDN_ENABLE_PINS) } {
+        lappend arg_list -pins "$::env(PDN_VERTICAL_LAYER)"
+    }
+
     define_pdn_grid \
         -name stdcell_grid \
         -starts_with POWER \
         -voltage_domain CORE \
-        -pins $::env(PDN_VERTICAL_LAYER)
+        {*}$arg_list
+
+    set arg_list [list]
+    append_if_equals arg_list PDN_EXTEND_TO "core_ring" -extend_to_core_ring
+    append_if_equals arg_list PDN_EXTEND_TO "boundary" -extend_to_boundary
 
     add_pdn_stripe \
         -grid stdcell_grid \
@@ -85,7 +114,8 @@ if { $::env(PDN_MULTILAYER) == 1 } {
         -pitch $::env(PDN_VPITCH) \
         -offset $::env(PDN_VOFFSET) \
         -spacing $::env(PDN_VSPACING) \
-        -starts_with POWER -extend_to_core_ring
+        -starts_with POWER \
+        {*}$arg_list
 }
 
 # Adds the standard cell rails if enabled.
@@ -105,21 +135,50 @@ if { $::env(PDN_ENABLE_RAILS) == 1 } {
 # Adds the core ring if enabled.
 if { $::env(PDN_CORE_RING) == 1 } {
     if { $::env(PDN_MULTILAYER) == 1 } {
+        set arg_list [list]
+        append_if_flag arg_list PDN_CORE_RING_ALLOW_OUT_OF_DIE -allow_out_of_die
+        append_if_flag arg_list PDN_CORE_RING_CONNECT_TO_PADS -connect_to_pads
+        append_if_equals arg_list PDN_EXTEND_TO "boundary" -extend_to_boundary
+
+        set pdn_core_vertical_layer $::env(PDN_VERTICAL_LAYER)
+        set pdn_core_horizontal_layer $::env(PDN_HORIZONTAL_LAYER)
+
+        if { [info exists ::env(PDN_CORE_VERTICAL_LAYER)] } {
+            set pdn_core_vertical_layer $::env(PDN_CORE_VERTICAL_LAYER)
+        }
+
+        if { [info exists ::env(PDN_CORE_HORIZONTAL_LAYER)] } {
+            set pdn_core_horizontal_layer $::env(PDN_CORE_HORIZONTAL_LAYER)
+        }
+
         add_pdn_ring \
-            -allow_out_of_die \
             -grid stdcell_grid \
-            -layers "$::env(PDN_VERTICAL_LAYER) $::env(PDN_HORIZONTAL_LAYER)" \
+            -layers "$pdn_core_vertical_layer $pdn_core_horizontal_layer" \
             -widths "$::env(PDN_CORE_RING_VWIDTH) $::env(PDN_CORE_RING_HWIDTH)" \
             -spacings "$::env(PDN_CORE_RING_VSPACING) $::env(PDN_CORE_RING_HSPACING)" \
-            -core_offset "$::env(PDN_CORE_RING_VOFFSET) $::env(PDN_CORE_RING_HOFFSET)"
+            -core_offset "$::env(PDN_CORE_RING_VOFFSET) $::env(PDN_CORE_RING_HOFFSET)" \
+            {*}$arg_list
+
+        if { [info exists ::env(PDN_CORE_VERTICAL_LAYER)] } {
+            add_pdn_connect \
+                -grid stdcell_grid \
+                -layers "$::env(PDN_CORE_VERTICAL_LAYER) $::env(PDN_HORIZONTAL_LAYER)"
+        }
+
+        if { [info exists ::env(PDN_CORE_HORIZONTAL_LAYER)] } {
+            add_pdn_connect \
+                -grid stdcell_grid \
+                -layers "$::env(PDN_CORE_HORIZONTAL_LAYER) $::env(PDN_VERTICAL_LAYER)"
+        }
+
+        if { [info exists ::env(PDN_CORE_VERTICAL_LAYER)] && [info exists ::env(PDN_CORE_HORIZONTAL_LAYER)] } {
+            add_pdn_connect \
+                -grid stdcell_grid \
+                -layers "$::env(PDN_CORE_VERTICAL_LAYER) $::env(PDN_CORE_HORIZONTAL_LAYER)"
+        }
+
     } else {
         throw APPLICATION "PDN_CORE_RING cannot be used when PDN_MULTILAYER is set to false."
-        # add_pdn_ring \
-        #     -grid stdcell_grid \
-        #     -layers "$::env(PDN_VERTICAL_LAYER)" \
-        #     -widths "$::env(PDN_CORE_RING_VWIDTH)" \
-        #     -spacings "$::env(PDN_CORE_RING_VSPACING)" \
-        #     -core_offset "$::env(PDN_CORE_RING_VOFFSET)"
     }
 }